+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Apply additional impulse when moving platform changes velocity to prevent
+  https://github.com/idanarye/bevy-tnua/issues/13.
 
 ## 0.2.2 - 2023-04-15
 ### Added