+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.3.0 - 2023-05-13
 ### Changed
 - [**BREAKING**] Removed `TnuaPlatformerBundle::new_with_config`. Users should
   use this instead: